Disables CMS toolbar and enable pause feature of cycle plugin
Currently there is a conflict between the jQuery versions
provided by django-cms and our own version. Right now I
don't think anyone is actually using the toolbar so I've
disabled it for now.

I've also enabled the cycle plugin's pause feature which
disables the rotation if the user is currently hovering
the container.