Fix failing CC payment in live-mode
The bridge communication contained the right data for the wrong
attribute when communicating the value in the initial token
generation request. This resulted in a currency-mismatch error at
the transaction-initialisation stage. "amount_int" should be used for
cent values and not "amount".

Details on this can be found on the PAYMILL blog:
https://blog.paymill.com/clarification-on-amount-int/