Consistent order of description and structure.
For trainings the order of the fields should be consistently set to
description, order and notes. We could also have simply changed the
order in the proposal form but that would have tried to change
expectations after the fact since the proposal edit form might no longer
be accessible at this time.